+++ include/nstd/collections.h
@@ -1,6 +1,7 @@
 #ifndef NSTD_COLLECTIONS_H_INCLUDED
 #define NSTD_COLLECTIONS_H_INCLUDED
 #include "collections/bit_mask.h"
+#include "collections/rc.h"
 #include "collections/vec.h"
 #include "nstd.h"
 #endif