+++ src/collections.rs
@@ -1,2 +1,3 @@
 pub mod bit_mask;
+pub mod rc;
 pub mod vec;